requirements.txt
Django==1.11.2
django-crispy-forms==1.6.1
